Limex: repaired VTKOutputObserver
It did not work any more in case the solution was to be printed only at specified time intervals.
This was due to an change in the interface ITimeIntegratorObserver.
The VTKOutputObserver is now implemented as an observer statically attached to the stages
  start  finalize  end.
It can be attached using attach_observer() without specifying the individual stages,
they are deduced automatically.